UI Update
Updated the UI, fixed the borders between objects finally, everything
looks as it should, and I added a text area that messages can be stored
in. I put this on the right of the frame, and then I moved the
observers to the left because most chats take place on the right hand
of the screen, so i thought this could be useful.